+++ heroku-deploy/src/main/java/com/heroku/sdk/deploy/WarApp.java
@@ -3,6 +3,7 @@ package com.heroku.sdk.deploy;
 import java.io.File;
 import java.io.IOException;
 import java.net.URL;
+import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
@@ -18,7 +19,7 @@ public class WarApp extends App {
   }
 
   public WarApp(String buildPackDesc, String name, File warFile, File webappRunnerJar, File rootDir, File targetDir) {
-    super(buildPackDesc, name, rootDir, targetDir);
+    super(buildPackDesc, name, rootDir, targetDir, new ArrayList<String>());
     this.warFile = warFile;
     this.webappRunnerJar = webappRunnerJar;
   }