+++ maven-plugin/src/it/build-config-test/verify.groovy
@@ -9,8 +9,12 @@ String appName = props["heroku.appName"]
 
 try {
     def log = FileUtils.fileRead(new File(basedir, "build.log"));
-    assert log.contains("Creating slug"), "did not create the slug"
-    assert log.contains("Uploading slug"), "did not upload the slug"
+    assert log.contains("Creating build"), "the build was not created"
+    assert log.contains("Uploading build"), "the build was not uploaded"
+    assert log.contains("remote:"), "the buildpacks were not run"
+    assert log.contains("Multipack app detected"), "Multipack not detected"
+    assert log.contains("Null app detected"), "null buildpack not detected"
+    assert log.contains("heroku-maven-plugin app detected"), "jvm-common buildpack not detected"
     assert log.contains("BUILD SUCCESS"), "the build was not successful"
 
     def process = "heroku run java -version -a${appName}".execute()